# internal/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Drakkar API",
    "version": "0.1.87",
    "description": "Drakkar NZB/Usenet VFS media manager REST API.\n\nAuthentication: Pass a session cookie (`drakkar_session`) or Bearer API token (`Authorization: Bearer <token>` or `X-Api-Key: <token>`) on all requests except public paths.",
    "contact": { "name": "Drakkar", "url": "https://github.com/drakkar-media/drakkar" }
  },
  "servers": [{ "url": "", "description": "Current host" }],
  "components": {
    "securitySchemes": {
      "BearerAuth": { "type": "http", "scheme": "bearer" },
      "ApiKeyHeader": { "type": "apiKey", "in": "header", "name": "X-Api-Key" },
      "CookieAuth": { "type": "apiKey", "in": "cookie", "name": "drakkar_session" }
    },
    "schemas": {
...
      "APIToken": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "userId": { "type": "integer" },
          "name": { "type": "string" },
          "createdAt": { "type": "string", "format": "date-time" },
          "lastUsedAt": { "type": "string", "format": "date-time" },
          "expiresAt": { "type": "string", "format": "date-time" }
        }
      }
    }
  },
  "security": [
    { "BearerAuth": [] },
    { "ApiKeyHeader": [] },
    { "CookieAuth": [] }
  ],
...
  "paths": {
...
    "/api/auth/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "Login with username and password",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["username", "password"],
                "properties": {
                  "username": { "type": "string" },
                  "password": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Login successful — sets session cookie" },
          "401": { "description": "Invalid credentials" }
        }
      }
    },
...
    "/api/auth/tokens": {
      "get": {
        "tags": ["Auth"],
        "summary": "List API tokens for current user",
        "responses": {
          "200": {
            "description": "Array of API tokens",
            "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/APIToken" } } } }
          }
        }
      },
      "post": {
        "tags": ["Auth"],
        "summary": "Create a new API token",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name"],
                "properties": {
                  "name": { "type": "string" },
                  "expiresAt": { "type": "string", "format": "date-time", "description": "Optional expiry (empty = no expiry)" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Created token — raw `token` field only returned here, not in list endpoint",
            "content": { "application/json": { "schema": { "allOf": [{ "$ref": "#/components/schemas/APIToken" }, { "type": "object", "properties": { "token": { "type": "string" } } }] } } }
          }
        }
      }
    },
    "/api/auth/tokens/{id}": {
      "delete": {
        "tags": ["Auth"],
        "summary": "Delete an API token",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": { "200": { "description": "Deleted" } }
      }
    },
    "/api/users": {
      "get": {
        "tags": ["Auth"],
        "summary": "List users",
        "responses": { "200": { "description": "Array of users" } }
      },
      "post": {
        "tags": ["Auth"],
        "summary": "Create user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["username", "password"],
                "properties": {
                  "username": { "type": "string" },
                  "password": { "type": "string" },
                  "role": { "type": "string", "default": "admin" }
                }
              }
            }
          }
        },
        "responses": { "200": { "description": "Created user" } }
      }
    },
    "/api/users/{id}/password": {
      "put": {
        "tags": ["Auth"],
        "summary": "Change user password",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "type": "object", "properties": { "password": { "type": "string" } } } } }
        },
        "responses": { "200": { "description": "Password changed" } }
      }
    },
...
  }
}